+++ app/entities/order/choices.py
@@ -1,3 +1,4 @@
+import enum
 from enum import Enum
 
 from entities.users.models_user import ClientModel, EmployeeModel
@@ -12,3 +13,12 @@ class OrderFilter(str, Enum):
 
     def invert(self) -> str:
         return self.client if self.value == self.employee else self.employee
+
+
+class StatusOrder(str, enum.Enum):
+    WAIT = 'W'
+    CANCEL = 'C'
+    PAID = 'PN'
+    SUCCESS = 'P'
+    PROC_RETURN = 'CR'
+    RETURN = 'R'
