+++ app/entities/order/schemas/schema_order.py
@@ -5,8 +5,7 @@ from pydantic import condecimal
 from entities.order.models_order import OrderModel
 from entities.service_name.models_service_name import ServiceNameModel
 from entities.users.models_user import ClientModel, EmployeeModel
-from models.choices import StatusOrder
-from entities.order.choices import OrderFilter
+from entities.order.choices import OrderFilter, StatusOrder
 from schemas.base_schema import PaginationSchema, BasePydanticSchema
 from entities.category.schemas_category import CategoryInSchema
 from core.utils.pagination import Pagination
